requirements.txt
aioredis==1.3.1
alembic==1.7.6
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
Authlib==0.15.5
autobahn==21.11.1
Automat==20.2.0
Babel==2.9.1
bcrypt==3.2.0
bidict==0.21.4
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
decouple==0.0.7
distro==1.6.0
Django==3.2.8
django-cors-headers==4.0.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.3
eventlet==0.33.0
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Compress==1.10.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
hiredis==2.0.0
httpagentparser==1.9.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.10.1
importlib-resources==5.4.0
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
ldap3==2.9.1
Mako==1.1.6
MarkupSafe==2.0.1
msgpack==1.0.3
paramiko==2.9.2
passlib==1.7.4
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
pgadmin4==6.4
Pillow==8.3.2
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyotp==2.6.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
pywin32==227
pywinpty==1.1.6
PyYAML==5.4.1
qrcode==7.3.1
redis==2.6.2
requests==2.25.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
speaklater3==1.4
SQLAlchemy==1.4.31
sqlparse==0.4.2
sshtunnel==0.4.0
texttable==1.6.4
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
websocket-client==0.59.0
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.7.0
zope.interface==5.4.0
